float test added; walk_d added

+++ tests/plot_rtree.py
@@ -11,7 +11,7 @@ if __name__ == '__main__':
   fig, ax = plt.subplots()
   ax.plot([0,700], [0,700])
   for x, x_len, y, y_len in s:
-    ax.add_patch(Rectangle((int(x), int(y)), int(x_len), int(y_len), facecolor='none', edgecolor='red'))
+    ax.add_patch(Rectangle((float(x), float(y)), float(x_len), float(y_len), facecolor='none', edgecolor='red'))
   plt.show()
